+++ src/daft-distributed/src/pipeline_node/shuffles/mod.rs
@@ -1,4 +1,4 @@
 pub mod gather;
 pub mod pre_shuffle_merge;
-pub mod shuffle;
+pub mod repartition;
 pub mod translate_shuffle;
